requirements.txt
Django~=3.2.8
python-dotenv~=0.19.0
django-tinymce~=3.3.0
Pillow~=8.3.2
django-admin-sortable2~=1.0.2